+++ src/main/java/net/geforcemods/securitycraft/ClientHandler.java
@@ -418,13 +418,13 @@ public class ClientHandler {
 				return mixWithReinforcedTintIfEnabled(grassTint);
 			}
 
-			return 0xFFFFFF;
+			return ConfigHandler.CLIENT.reinforcedBlockTintColor.get();
 		}, SCContent.REINFORCED_GRASS_BLOCK.get());
 		event.getBlockColors().register((state, world, pos, tintIndex) -> {
 			if (tintIndex == 1)
 				return world != null && pos != null ? BiomeColors.getAverageWaterColor(world, pos) : -1;
 
-			return 0xFFFFFF;
+			return ConfigHandler.CLIENT.reinforcedBlockTintColor.get();
 		}, SCContent.REINFORCED_WATER_CAULDRON.get());
 	}
 
@@ -461,7 +461,7 @@ public class ClientHandler {
 				return mixWithReinforcedTintIfEnabled(grassTint);
 			}
 
-			return 0xFFFFFF;
+			return ConfigHandler.CLIENT.reinforcedBlockTintColor.get();
 		}, SCContent.REINFORCED_GRASS_BLOCK.get());
 		blocksWithReinforcedTint = null;
 		blocksWithCustomTint = null;