+++ Changelog.md
@@ -1,6 +1,7 @@
 --------------------------Changelog for v1.9.5 of SecurityCraft--------------------------
 
 - New: The reinforced block tint color can now be changed with a config setting
+- Fix: The Reinforced Water Cauldron and Reinforced Grass Block don't have a reinforced tint
 
 --------------------------Changelog for v1.9.4 of SecurityCraft--------------------------